# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pakperk API",
    "description": "Stable v1 paper-reading API with optional OIDC-backed account profiles. Library and comment routes remain feature-gated until their production phases are complete.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/feed": {
      "get": {
        "tags": [
          "routes::feed"
        ],
        "operationId": "feed",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "arXiv category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque page cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "description": "Entity tag returned by an earlier response for the same feed query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper summary page",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public chronological-feed cache policy"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque validator for this feed query and representation"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedPageSchema"
                }
              }
            }
          },
          "304": {
            "description": "Feed representation has not changed; response body is empty",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "Public chronological-feed cache policy"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Current opaque feed validator"
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
...
    "/v1/papers/{paper_id}/connections": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "connections",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}/introduction": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "introduction",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper introduction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntroductionSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapabilitiesSchema": {
        "type": "object",
        "required": [
          "metadata",
          "introduction",
          "chat",
          "connections"
        ],
        "properties": {
          "chat": {
            "type": "boolean"
          },
          "connections": {
            "type": "boolean"
          },
          "introduction": {
            "type": "boolean"
          },
          "metadata": {
            "type": "boolean"
          }
        }
      },
...
      "ConnectionReferenceSchema": {
        "type": "object",
        "required": [
          "ordinal",
          "raw_text",
          "resolved",
          "resolution_status"
        ],
        "properties": {
          "ordinal": {
            "type": "integer",
            "minimum": 0
          },
          "paper_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "raw_text": {
            "type": "string"
          },
          "resolution_status": {
            "$ref": "#/components/schemas/ReferenceResolutionStatusSchema"
          },
          "resolved": {
            "type": "boolean"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ConnectionsResponseSchema": {
        "type": "object",
        "required": [
          "paper_id",
          "generation",
          "ready",
          "key_connections",
          "references"
        ],
        "properties": {
          "generation": {
            "type": "integer",
            "format": "int32"
          },
          "key_connections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyConnectionSchema"
            }
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "ready": {
            "type": "boolean"
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectionReferenceSchema"
            }
          }
        }
      },
      "ErrorBodySchema": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code.",
            "example": "PAPER_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "The requested paper was not found."
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ErrorEnvelopeSchema": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodySchema"
          }
        }
      },
...
      "FeedPageSchema": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaperSummarySchema"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "IntroductionDetectionSchema": {
        "type": "object",
        "required": [
          "confidence",
          "used_fallback"
        ],
        "properties": {
          "confidence": {
            "type": "number",
            "format": "float"
          },
          "used_fallback": {
            "type": "boolean"
          }
        }
      },
...
      "IntroductionSchema": {
        "type": "object",
        "required": [
          "paper_id",
          "generation",
          "paragraphs",
          "detection",
          "original_pdf_url"
        ],
        "properties": {
          "detection": {
            "$ref": "#/components/schemas/IntroductionDetectionSchema"
          },
          "generation": {
            "type": "integer",
            "format": "int32"
          },
          "heading": {
            "type": [
              "string",
              "null"
            ]
          },
          "original_pdf_url": {
            "type": "string"
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "paragraphs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntroductionParagraphSchema"
            }
          }
        }
      },
      "KeyConnectionSchema": {
        "type": "object",
        "required": [
          "reference_id",
          "paper_id",
          "arxiv_id",
          "title",
          "authors",
          "relation_type",
          "summary",
          "confidence"
        ],
        "properties": {
          "arxiv_id": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence": {
            "type": "number",
            "format": "float"
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "reference_id": {
            "type": "string",
            "format": "uuid"
          },
          "relation_type": {
            "$ref": "#/components/schemas/RelationTypeSchema"
          },
          "summary": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "PaperSummarySchema": {
        "type": "object",
        "required": [
          "paper_id",
          "arxiv_id",
          "title",
          "abstract",
          "authors",
          "primary_category",
          "categories",
          "published_at",
          "updated_at",
          "abs_url",
          "pdf_url",
          "capabilities"
        ],
        "properties": {
          "abs_url": {
            "type": "string"
          },
          "abstract": {
            "type": "string"
          },
          "arxiv_id": {
            "type": "string",
            "example": "2401.12345v2"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesSchema"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "pdf_url": {
            "type": "string"
          },
          "primary_category": {
            "type": "string"
          },
          "published_at": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oidcBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC access token validated against the configured issuer, audience, algorithm, and JWKS."
      }
    }
  },
...
}